+++ src/py42/_internal/clients/legal_hold.py
@@ -6,14 +6,12 @@ import py42.settings as settings
 
 
 class LegalHoldClient(BaseAuthorityClient):
-    def create_legal_hold_policy(self, name, policy=None):
+    def create_policy(self, name, policy=None):
         uri = u"/api/v4/legal-hold-policy/create"
         data = {u"name": name, u"policy": policy}
         return self._default_session.post(uri, data=json.dumps(data))
 
-    def create_legal_hold(
-        self, name, hold_policy_uid, description=None, notes=None, hold_ext_ref=None
-    ):
+    def create_matter(self, name, hold_policy_uid, description=None, notes=None, hold_ext_ref=None):
         uri = u"/api/LegalHold"
         data = {
             u"name": name,
@@ -24,16 +22,16 @@ class LegalHoldClient(BaseAuthorityClient):
         }
         return self._default_session.post(uri, data=json.dumps(data))
 
-    def get_legal_hold_policy_by_uid(self, legal_hold_policy_uid):
+    def get_policy_by_uid(self, legal_hold_policy_uid):
         uri = u"/api/v4/legal-hold-policy/view"
         params = {u"legalHoldPolicyUid": legal_hold_policy_uid}
         return self._v3_required_session.get(uri, params=params)
 
-    def get_all_legal_hold_policies(self):
+    def get_all_policies(self):
         uri = u"/api/v4/legal-hold-policy/list"
         return self._v3_required_session.get(uri)
 
-    def get_legal_hold_by_uid(self, legal_hold_uid):
+    def get_matter_by_uid(self, legal_hold_uid):
         uri = u"/api/LegalHold/{0}".format(legal_hold_uid)
         return self._default_session.get(uri)
 
@@ -57,7 +55,7 @@ class LegalHoldClient(BaseAuthorityClient):
         }
         return self._default_session.get(uri, params=params)
 
-    def get_legal_holds(
+    def get_all_matters(
         self, creator_user_uid=None, active_state=u"ACTIVE", name=None, hold_ext_ref=None
     ):
         return get_all_pages(
@@ -92,7 +90,7 @@ class LegalHoldClient(BaseAuthorityClient):
         uri = u"/api/LegalHoldMembership"
         return self._default_session.get(uri, params=params)
 
-    def get_legal_hold_memberships(
+    def get_all_matter_custodians(
         self,
         legal_hold_membership_uid=None,
         legal_hold_uid=None,
@@ -111,21 +109,21 @@ class LegalHoldClient(BaseAuthorityClient):
             active_state=active_state,
         )
 
-    def add_user_to_legal_hold(self, user_uid, legal_hold_uid):
+    def add_to_matter(self, user_uid, legal_hold_uid):
         uri = u"/api/LegalHoldMembership"
         data = {u"legalHoldUid": legal_hold_uid, u"userUid": user_uid}
         return self._default_session.post(uri, data=json.dumps(data))
 
-    def remove_user_from_legal_hold(self, legal_hold_membership_uid):
+    def remove_from_matter(self, legal_hold_membership_uid):
         uri = u"/api/LegalHoldMembershipDeactivation"
         data = {u"legalHoldMembershipUid": legal_hold_membership_uid}
         return self._default_session.post(uri, data=json.dumps(data))
 
-    def deactivate_legal_hold(self, legal_hold_uid):
+    def deactivate_matter(self, legal_hold_uid):
         uri = u"/api/v4/legal-hold-deactivation/update"
         data = {u"legalHoldUid": legal_hold_uid}
         return self._v3_required_session.post(uri, data=json.dumps(data))
 
-    def reactivate_legal_hold(self, legal_hold_uid):
+    def reactivate_matter(self, legal_hold_uid):
         uri = u"/api/LegalHoldReactivation/{0}".format(legal_hold_uid)
         return self._default_session.put(uri)